+++ packages/gatsby-codemods/src/transforms/__testfixtures__/global-graphql-calls/import-default.output.js
@@ -7,4 +7,4 @@ export const query = graphql`
       prefix
     }
   }
-`
+`;