+++ packages/gatsby-codemods/src/transforms/__testfixtures__/global-graphql-calls/import-namespace.output.js
@@ -1,5 +1,5 @@
 /* eslint-disable */
-import * as Gatsby from "gatsby"
+import * as Gatsby from "gatsby";
 
 export const query = Gatsby.graphql`
   query {
@@ -7,4 +7,4 @@ export const query = Gatsby.graphql`
       prefix
     }
   }
-`
+`;