+++ packages/gatsby-codemods/src/transforms/__testfixtures__/global-graphql-calls/no-import-esm.input.js
@@ -1,9 +1,8 @@
 /* eslint-disable */
-// TODO: update codemod to make this test pass
 export const query = graphql`
   query {
     allSitePages {
       prefix
     }
   }
-`
+`;