+++ packages/gatsby-plugin-flow/src/gatsby-node.js
@@ -1,0 +1,7 @@
+function onCreateBabelConfig({ actions }, pluginOptions) {
+  actions.setBabelPreset({
+    name: `@babel/preset-flow`,
+  })
+}
+
+exports.onCreateBabelConfig = onCreateBabelConfig