+++ packages/gatsby-source-wordpress/src/gatsby-node.js
@@ -135,6 +135,8 @@ exports.sourceNodes = async (
     searchAndReplaceContentUrls,
   })
 
+  entities = normalize.mapPolylangTranslations(entities)
+
   // apply custom normalizer
   if (typeof _normalizer === `function`) {
     entities = _normalizer({