+++ packages/gatsby/src/redux/actions.js
@@ -7,6 +7,7 @@ const { stripIndent } = require(`common-tags`)
 const report = require(`gatsby-cli/lib/reporter`)
 const path = require(`path`)
 const fs = require(`fs`)
+const url = require(`url`)
 const kebabHash = require(`kebab-hash`)
 const { hasNodeChanged, getNode } = require(`./index`)
 const { trackInlineObjectsInRootNode } = require(`../schema/node-tracking`)
@@ -893,7 +894,9 @@ actions.replaceWebpackConfig = (config: Object, plugin?: ?Plugin = null) => {
  * @param {Object} config An options object in the shape of a normal babelrc javascript object
  * @example
  * setBabelOptions({
- *   sourceMaps: `inline`,
+ *   options: {
+ *     sourceMaps: `inline`,
+ *   }
  * })
  */
 actions.setBabelOptions = (options: Object, plugin?: ?Plugin = null) => {
@@ -1097,13 +1100,19 @@ actions.createRedirect = ({
     pathPrefix = store.getState().config.pathPrefix
   }
 
+  // Parse urls to get their protocols
+  // url.parse will not cover protocol-relative urls so do a separate check for those
+  const parsed = url.parse(toPath)
+  const isRelativeProtocol = toPath.startsWith(`//`)
+  const toPathPrefix = parsed.protocol != null || isRelativeProtocol ? `` : pathPrefix
+
   return {
     type: `CREATE_REDIRECT`,
     payload: {
       fromPath: `${pathPrefix}${fromPath}`,
       isPermanent,
       redirectInBrowser,
-      toPath: `${pathPrefix}${toPath}`,
+      toPath: `${toPathPrefix}${toPath}`,
       ...rest,
     },
   }
@@ -1134,11 +1143,11 @@ actions.addThirdPartySchema = (
 }
 
 /**
- * All defined actions.
+ * All action creators wrapped with a dispatch.
  */
 exports.actions = actions
 
 /**
- * All action creators wrapped with a dispatch.
+ * All action creators wrapped with a dispatch. - *DEPRECATED*
  */
 exports.boundActionCreators = bindActionCreators(actions, store.dispatch)