+++ www/src/components/sidebar/sticky-responsive-sidebar.js
@@ -1,10 +1,13 @@
-import React, { Component, Fragment } from "react"
+import React, { Component } from "react"
 
 import Sidebar from "./sidebar"
 import ScrollSyncSidebar from "./scroll-sync-sidebar"
 import ChevronSvg from "./chevron-svg"
 import presets, { colors } from "../../utils/presets"
 import { rhythm } from "../../utils/typography"
+import ScrollPositionProvider, {
+  ScrollPositionConsumer,
+} from "./scrollbar-position-provider"
 
 class StickyResponsiveSidebar extends Component {
   constructor(props) {
@@ -23,16 +26,20 @@ class StickyResponsiveSidebar extends Component {
 
   render() {
     const { open } = this.state
-    const SidebarComponent = this.props.enableScrollSync
-      ? ScrollSyncSidebar
-      : Sidebar
+    const {
+      enableScrollSync,
+      location: { pathname },
+    } = this.props
+    const SidebarComponent = enableScrollSync ? ScrollSyncSidebar : Sidebar
 
-    const iconOffset = open ? 8 : -4
+    const iconOffset = open ? 5 : -5
     const menuOpacity = open ? 1 : 0
     const menuOffset = open ? 0 : rhythm(10)
 
+    const sidebarType = pathname.split(`/`)[1]
+
     return (
-      <Fragment>
+      <ScrollPositionProvider>
         <div
           css={{
             ...styles.sidebarScrollContainer,
@@ -46,36 +53,45 @@ class StickyResponsiveSidebar extends Component {
               transform: `translateX(-${menuOffset})`,
             }}
           >
-            <SidebarComponent
-              closeSidebar={this._closeSidebar}
-              {...this.props}
-            />
+            <ScrollPositionConsumer>
+              {({ positions, onPositionChange }) => (
+                <SidebarComponent
+                  position={positions[sidebarType]}
+                  onPositionChange={onPositionChange}
+                  closeSidebar={this._closeSidebar}
+                  {...this.props}
+                />
+              )}
+            </ScrollPositionConsumer>
           </div>
         </div>
         <div
           css={{ ...styles.sidebarToggleButton }}
           onClick={this._openSidebar}
           role="button"
+          aria-label="Show Secondary Navigation"
+          aria-controls="SecondaryNavigation"
+          aria-expanded={open ? "true" : "false"}
           tabIndex={0}
         >
           <div css={{ ...styles.sidebarToggleButtonInner }}>
             <ChevronSvg
               size={15}
               cssProps={{
-                transform: `translate(2px, ${iconOffset}px) rotate(180deg)`,
+                transform: `translate(${iconOffset}px, 5px) rotate(90deg)`,
                 transition: `transform 0.2s ease`,
               }}
             />
             <ChevronSvg
               size={15}
               cssProps={{
-                transform: `translate(2px, ${0 - iconOffset}px)`,
+                transform: `translate(${5 - iconOffset}px, -5px) rotate(270deg)`,
                 transition: `transform 0.2s ease`,
               }}
             />
           </div>
         </div>
-      </Fragment>
+      </ScrollPositionProvider>
     )
   }
 }
@@ -94,13 +110,11 @@ const styles = {
     width: 320,
     zIndex: 10,
     [presets.Tablet]: {
-      height: `calc(100vh - ${presets.headerHeight} - ${
-        presets.bannerHeight
-      } + 1px)`,
+      height: `calc(100vh - ${presets.headerHeight} - ${presets.bannerHeight})`,
       maxWidth: `none`,
       opacity: `1 !important`,
       pointerEvents: `auto`,
-      top: `calc(${presets.headerHeight} + ${presets.bannerHeight} - 1px)`,
+      top: `calc(${presets.headerHeight} + ${presets.bannerHeight})`,
       width: rhythm(10),
     },
     [presets.Desktop]: {