+++ packages/gatsby-plugin-sass/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.1"></a>
+
+## [2.0.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-sass/compare/gatsby-plugin-sass@2.0.0-rc.3...gatsby-plugin-sass@2.0.1) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-plugin-sass
+
+<a name="2.0.0-rc.3"></a>
+
+# [2.0.0-rc.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-sass/compare/gatsby-plugin-sass@2.0.0-rc.2...gatsby-plugin-sass@2.0.0-rc.3) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-plugin-sass
+
+<a name="2.0.0-rc.2"></a>
+
+# [2.0.0-rc.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-sass/compare/gatsby-plugin-sass@2.0.0-rc.1...gatsby-plugin-sass@2.0.0-rc.2) (2018-09-04)
+
+**Note:** Version bump only for package gatsby-plugin-sass
+
+<a name="2.0.0-rc.1"></a>
+
+# [2.0.0-rc.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-sass/compare/gatsby-plugin-sass@2.0.0-rc.0...gatsby-plugin-sass@2.0.0-rc.1) (2018-08-29)
+
+**Note:** Version bump only for package gatsby-plugin-sass
+
+<a name="2.0.0-rc.0"></a>
+
+# [2.0.0-rc.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-sass/compare/gatsby-plugin-sass@2.0.0-beta.10...gatsby-plugin-sass@2.0.0-rc.0) (2018-08-21)
+
+**Note:** Version bump only for package gatsby-plugin-sass
+
 <a name="2.0.0-beta.10"></a>
 
 # [2.0.0-beta.10](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-sass/compare/gatsby-plugin-sass@2.0.0-beta.9...gatsby-plugin-sass@2.0.0-beta.10) (2018-08-08)