+++ packages/gatsby-react-router-scroll/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.0"></a>
+
+# [2.0.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-react-router-scroll/compare/gatsby-react-router-scroll@2.0.0-rc.2...gatsby-react-router-scroll@2.0.0) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-react-router-scroll
+
+<a name="2.0.0-rc.2"></a>
+
+# [2.0.0-rc.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-react-router-scroll/compare/gatsby-react-router-scroll@2.0.0-rc.1...gatsby-react-router-scroll@2.0.0-rc.2) (2018-08-31)
+
+**Note:** Version bump only for package gatsby-react-router-scroll
+
+<a name="2.0.0-rc.1"></a>
+
+# [2.0.0-rc.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-react-router-scroll/compare/gatsby-react-router-scroll@2.0.0-rc.0...gatsby-react-router-scroll@2.0.0-rc.1) (2018-08-29)
+
+**Note:** Version bump only for package gatsby-react-router-scroll
+
+<a name="2.0.0-rc.0"></a>
+
+# [2.0.0-rc.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-react-router-scroll/compare/gatsby-react-router-scroll@2.0.0-beta.9...gatsby-react-router-scroll@2.0.0-rc.0) (2018-08-21)
+
+**Note:** Version bump only for package gatsby-react-router-scroll
+
+<a name="2.0.0-beta.9"></a>
+
+# [2.0.0-beta.9](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-react-router-scroll/compare/gatsby-react-router-scroll@2.0.0-beta.8...gatsby-react-router-scroll@2.0.0-beta.9) (2018-08-21)
+
+**Note:** Version bump only for package gatsby-react-router-scroll
+
 <a name="2.0.0-beta.8"></a>
 
 # [2.0.0-beta.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-react-router-scroll/compare/gatsby-react-router-scroll@2.0.0-beta.6...gatsby-react-router-scroll@2.0.0-beta.8) (2018-08-07)