+++ packages/gatsby-transformer-sharp/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.3"></a>
+
+## [2.1.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-sharp/compare/gatsby-transformer-sharp@2.1.2...gatsby-transformer-sharp@2.1.3) (2018-09-28)
+
+### Features
+
+- provide fragments without gatsby-image ([#8459](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-sharp/issues/8459)) ([3389b89](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-sharp/commit/3389b89))
+
+<a name="2.1.2"></a>
+
+## [2.1.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-sharp/compare/gatsby-transformer-sharp@2.1.1-rc.3...gatsby-transformer-sharp@2.1.2) (2018-09-24)
+
+### Features
+
+- **gatsby-transformer-sharp:** expose sizes argument for fluid query ([#8466](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-sharp/issues/8466)) ([efe95a4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-sharp/commit/efe95a4))
+
+<a name="2.1.1"></a>
+
+## [2.1.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-sharp/compare/gatsby-transformer-sharp@2.1.1-rc.3...gatsby-transformer-sharp@2.1.1) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-transformer-sharp
+
+<a name="2.1.1-rc.3"></a>
+
+## [2.1.1-rc.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-sharp/compare/gatsby-transformer-sharp@2.1.1-rc.2...gatsby-transformer-sharp@2.1.1-rc.3) (2018-09-07)
+
+**Note:** Version bump only for package gatsby-transformer-sharp
+
+<a name="2.1.1-rc.2"></a>
+
+## [2.1.1-rc.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-sharp/compare/gatsby-transformer-sharp@2.1.1-rc.1...gatsby-transformer-sharp@2.1.1-rc.2) (2018-08-29)
+
+**Note:** Version bump only for package gatsby-transformer-sharp
+
+<a name="2.1.1-rc.1"></a>
+
+## [2.1.1-rc.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-sharp/compare/gatsby-transformer-sharp@2.1.1-rc.0...gatsby-transformer-sharp@2.1.1-rc.1) (2018-08-29)
+
+**Note:** Version bump only for package gatsby-transformer-sharp
+
+<a name="2.1.1-rc.0"></a>
+
+## [2.1.1-rc.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-sharp/compare/gatsby-transformer-sharp@2.1.1-beta.7...gatsby-transformer-sharp@2.1.1-rc.0) (2018-08-21)
+
+**Note:** Version bump only for package gatsby-transformer-sharp
+
 <a name="2.1.1-beta.7"></a>
 
 ## [2.1.1-beta.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-sharp/compare/gatsby-transformer-sharp@2.1.1-beta.6...gatsby-transformer-sharp@2.1.1-beta.7) (2018-08-13)